requirements.txt
flask==3.1.1
supabase==2.13.0
bcrypt==4.3.0
resend==2.7.0
requests==2.32.3
nba_api==1.10.2
python-dotenv==1.1.0
pytest==8.3.5
